+++ weblate/settings_docker.py
@@ -704,6 +704,9 @@ VCS_ALLOW_SCHEMES = set(get_env_list("WEBLATE_VCS_ALLOW_SCHEMES", ["https", "ssh
 
 # Email registration filter
 REGISTRATION_EMAIL_MATCH = get_env_str("WEBLATE_REGISTRATION_EMAIL_MATCH", ".*")
+REGISTRATION_ALLOW_DISPOSABLE_EMAILS = get_env_bool(
+    "WEBLATE_REGISTRATION_ALLOW_DISPOSABLE_EMAILS", False
+)
 
 private_commit_email_template_str = get_env_str("WEBLATE_PRIVATE_COMMIT_EMAIL_TEMPLATE")
 if private_commit_email_template_str is not None: