+++ weblate/utils/validators.py
@@ -230,7 +230,9 @@ class EmailValidator(EmailValidatorDjango):
                 gettext("Invalid e-mail address: {}").format(error)
             ) from error
 
-        if address.domain in blocklist:
+        if address.domain in blocklist and not getattr(
+            settings, "REGISTRATION_ALLOW_DISPOSABLE_EMAILS", False
+        ):
             raise ValidationError(gettext("Disposable e-mail domains are disallowed."))